+++ lib/printly.dart
@@ -7,4 +7,8 @@
 /// ```
 library;
 
+export 'package:permission_handler/permission_handler.dart'
+    show PermissionStatus;
+
+export 'src/bluetooth/bluetooth_adapter_state.dart';
 export 'src/printly_base.dart';